Added EmailService

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -5,19 +5,22 @@ import net.engineeringdigest.journalApp.entity.JournalEntry;
 import net.engineeringdigest.journalApp.entity.User;
 import net.engineeringdigest.journalApp.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
 import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
-@Component
+@Service
 public class JournalEntryService{
 
     @Autowired
     private JournalEntryRepository journalEntryRepository;
+
     @Autowired
     private UserService userService;
 
@@ -30,7 +33,7 @@ public class JournalEntryService{
             user.getJournalEntries().add(saved);
             userService.saveNewUser(user);
         }catch (Exception e){
-            throw  new RuntimeException("An error occured while saving the entry", e);
+            throw  new RuntimeException("An error occurred while saving the entry", e);
         }
     }
 
@@ -38,18 +41,23 @@ public class JournalEntryService{
         return journalEntryRepository.findAll();
     }
 
-    public Optional<JournalEntry> getById(ObjectId id){
+    public Optional<JournalEntry> findById(ObjectId id){
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveNewUser(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            throw new RuntimeException("An error occurred while deleting the entry.", e);
+        }
+        return removed;
     }
-
-//    public List<JournalEntry> findByUserName(String username){
-//    }
-
 }
